Drop redundant back/forward buttons
The path breadcrumb has back/forward buttons to its left which are
essentially a redundant reimplementation of the browser back/forward
buttons.  Let's drop that.

Dropping these will also clean up a not inconsiderable amount of the
amount of state we're passing around between components.

Rename the simplified breadcrumb component to .tsx
