+++ test/check-application
@@ -195,25 +195,25 @@ class TestNavigator(testlib.MachineCase):
         b.wait_text(".last-breadcrumb-button", "home")
         b.wait_visible("[data-item='admin']")
         # navigate back
-        b.click("#navigate-back")
+        b.eval_js("window.history.back()")
         b.wait_in_text("#sidebar-card-header", "newdir2")
         b.wait_not_present("[data-item='admin']")
-        b.click("#navigate-back")
+        b.eval_js("window.history.back()")
         b.wait_in_text("#sidebar-card-header", "newdir")
         b.wait_visible("[data-item='newdir2']")
-        b.click("#navigate-back")
+        b.eval_js("window.history.back()")
         b.wait_in_text("#sidebar-card-header", "admin")
         b.wait_visible("[data-item='newdir']")
         # navigate forward
-        b.click("#navigate-forward")
+        b.eval_js("window.history.forward()")
         b.wait_in_text("#sidebar-card-header", "newdir")
         b.wait_not_present("[data-item='admin']")
         b.wait_text(".last-breadcrumb-button", "newdir")
-        b.click("#navigate-forward")
+        b.eval_js("window.history.forward()")
         b.wait_in_text("#sidebar-card-header", "newdir2")
         b.wait_not_present("[data-item='newdir']")
         b.wait_text(".last-breadcrumb-button", "newdir2")
-        b.click("#navigate-forward")
+        b.eval_js("window.history.forward()")
         b.click("#navigator-card-body")  # unselect
         b.wait_in_text("#sidebar-card-header", "home")
         b.wait_not_present("[data-item='newdir']")
